+++ tests/test_fft.py
@@ -47,7 +47,8 @@ class MySTFT(nn.Module):
 
 INPUT_AND_MODELS = []
 
-if not tract_version_lower_than("0.20.0"):
+# disabled for now as tract have a feature gate on this
+if not tract_version_lower_than("0.20.0") and False:
     INPUT_AND_MODELS += [(torch.FloatTensor([[0, 1], [2, 3]]), MyFFT())]
     INPUT_AND_MODELS += [
         (